requirements.txt
gradio
python-dotenv
beautifulsoup4==4.12.3
google-generativeai
python==3.11.10